pr7/requirements.txt
django
pillow 
django-allauth
django-unfold
PYJWT 
requests 
cryptography